requirements.txt
annotated-types==0.7.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
es_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.8.0/es_core_news_lg-3.8.0-py3-none-any.whl#sha256=7c6c212715a12f31aacde3361754436945ff7376fb24cde57d0c277c9c9b050b
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.2.6
packaging==25.0
pillow==11.3.0
preshed==3.0.10
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyPDF2==3.0.1
python-docx==1.2.0
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
smart_open==7.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
